requirements.txt
# --- Core Deep Learning ---
torch>=2.1.0        # Updated for better SDPA (Scaled Dot Product Attention) speed
torchvision>=0.16.0
numpy>=1.24.0

# --- The Hugging Face Ecosystem (Critical for QLoRA) ---
transformers>=4.38.0 
peft>=0.10.0          
accelerate>=0.27.0    # Required for device_map="auto" and offloading
bitsandbytes==0.42.0  # The engine for 4-bit NF4 quantization
datasets>=2.14.0 


# --- Legacy / Baselines ---
# Using HF Transformers now, but keeping this if it needs to run 'baseline_eval.py'
# segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# --- Image Processing & Data ---
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.10.0        

# --- Experiment Tracking & Config ---
wandb>=0.16.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# --- Visualization & Analysis ---
matplotlib>=3.8.0
pandas>=2.0.0
seaborn>=0.13.0
tqdm>=4.66.0

# --- Dev Tools ---
jupyter>=1.0.0
ipykernel>=6.25.0